% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhw_plot.R
\name{plot_rasters_squish_outliers}
\alias{plot_rasters_squish_outliers}
\title{palette_name must be 
   https://dieghernan.github.io/tidyterra/reference/grass_db.html, 
   and you can preview the colors on 
   https://grass.osgeo.org/grass83/manuals/r.colors.html (but get the name
   from the tidyterra reference)
   
   
plot terra rasters, but manage outliers}
\usage{
plot_rasters_squish_outliers(
  raster_list,
  title = "",
  subtitle = "",
  scale_label = "",
  cut_percent = 0,
  max_threshold_value = NA,
  min_threshold_value = NA,
  break_width = NA,
  palette_name = "bgyr"
)
}
\arguments{
\item{raster_list}{terra spat raster list (or stack?)}

\item{title}{character title of the plot, default is empty title}

\item{scale_label}{character, label to print above the legend/scale}

\item{cut_percent}{numeric percentile cut-off, see percentile_cutoff_value function for descriptoin, but 1 
will cut top 1 percent, 0.25 will cut to 0.25\% off and display as top color (e.g. red)}

\item{max_threshold_value}{numeric optional numeric cut off the range display.  if this is sent, then cut_percent is ignored}

\item{min_threshold_value}{numeric optional numeric cut off the bottom of the range.}

\item{break_width}{numeric, default NA.  If this is set, then uses the non-continous bars and this is the width of those}

\item{palette_name}{string this must be a valid palette from}
}
\value{
ggplot object, one map for each item in raster_list with single legend
}
\description{
given a list of terra rasters to plot, use heat gradient to display values, 
but cut off the top X percent and replace those values with the top color (red in this case)
in order to handle extreme outliers.  Currently this does _not_ print the max range (outliers) in the legend
}
