% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhw_summary.R
\name{summary_by_decades_raster}
\alias{summary_by_decades_raster}
\title{summarize a metric grouped by decades 2040 2050 2060}
\usage{
summary_by_decades_raster(
  mhwdb_conn,
  mhw_table = "arise10_decade_metrics",
  mhw_metric = "int_mean",
  sql_function = "avg",
  decades = c(2040, 2050, 2060),
  ensemble_list_string = NA,
  crs = "EPSG:4087"
)
}
\arguments{
\item{mhw_table}{name of the table to use}

\item{mhw_metric}{character name of metric ( column ) in the table to summarize: mhw_dur, int_mean, etc}

\item{decades}{vector of integers, the start year of each 10 year period, can overlap}

\item{ensemble_list_string}{= NA; optional character string.  If NA, then it's ignored and all ensembles are included. 
if it's string that  is a list of ensembles to include in format "006,007,008"  only those with matching 
ensemble will be included.  
ensembles have leading zeros, are 3 digits and this must be a separated list with leading zeros}

\item{crs}{default 'EPSG:4087', character valid coordinate reference system}

\item{sqlfun}{= avg; character aggregate function for duckdb, for example avg, count or median, 
see https://duckdb.org/docs/sql/functions/aggregates,}
}
\value{
list of terra rasters of summary statistics, one per decade in decades vector
}
\description{
runs summary and creates raster for a 10 year periods for given list of 
starting years. uses the summary_by_decades_raster() function and sends 
params to that, using the decade start year as the 'group id' value
}
